new/requirements.txt
# requirements.txt
# 核心依赖
requests==2.32.3
fastapi==0.116.0
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0

# Cython编译依赖
cython==3.1.2
setuptools==75.8.0
numpy==2.2.5

aiohttp==3.10.5
zhdate==0.1
